// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      courses: {
        Row: {
          category: string
          created_at: string
          description: string
          difficulty_level: string
          duration: string
          id: string
          image_url: string | null
          lessons_count: number
          lottie_url: string | null
          modules_count: number
          short_description: string
          slug: string
          target_audience: string[]
          title: string
          updated_at: string
          what_you_learn: string[]
        }
        Insert: {
          category: string
          created_at?: string
          description: string
          difficulty_level: string
          duration: string
          id?: string
          image_url?: string | null
          lessons_count: number
          lottie_url?: string | null
          modules_count: number
          short_description: string
          slug: string
          target_audience: string[]
          title: string
          updated_at?: string
          what_you_learn: string[]
        }
        Update: {
          category?: string
          created_at?: string
          description?: string
          difficulty_level?: string
          duration?: string
          id?: string
          image_url?: string | null
          lessons_count?: number
          lottie_url?: string | null
          modules_count?: number
          short_description?: string
          slug?: string
          target_audience?: string[]
          title?: string
          updated_at?: string
          what_you_learn?: string[]
        }
        Relationships: []
      }
      newsletter_subscriptions: {
        Row: {
          created_at: string
          email: string
          id: string
        }
        Insert: {
          created_at?: string
          email: string
          id?: string
        }
        Update: {
          created_at?: string
          email?: string
          id?: string
        }
        Relationships: []
      }
      testimonials: {
        Row: {
          course: string
          created_at: string
          id: string
          image_url: string | null
          name: string
          outcome: string | null
          rating: number
          role: string
          testimonial: string
        }
        Insert: {
          course: string
          created_at?: string
          id?: string
          image_url?: string | null
          name: string
          outcome?: string | null
          rating: number
          role: string
          testimonial: string
        }
        Update: {
          course?: string
          created_at?: string
          id?: string
          image_url?: string | null
          name?: string
          outcome?: string | null
          rating?: number
          role?: string
          testimonial?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never